requirements.txt
alembic==0.8.6
aniso8601==1.1.0
blinker==1.4
cov-core==1.15.0
coverage==3.7.1
cssmin==0.2.0
flake8==2.4.0
Flask==0.10.1
Flask-Assets==0.10
Flask-Cache==0.13.1
Flask-DebugToolbar==0.10.0
Flask-Login==0.2.11
Flask-Migrate==1.8.0
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
Flask-WTF==0.11
itsdangerous==0.24
Jinja2==2.8
jsmin==2.1.1
Mako==1.0.4
MarkupSafe==0.23
mccabe==0.3
nose==1.3.7
pep8==1.5.7
pushould-python==0.0.3
py==1.4.31
pyflakes==0.8.1
pytest==2.7.0
pytest-cov==1.8.1
python-dateutil==2.5.2
python-editor==1.0
pytz==2016.3
requests==2.9.1
six==1.10.0
SQLAlchemy==1.0.12
webassets==0.11.1
Werkzeug==0.11.8
WTForms==2.1
